post-service/requirements.txt
Django==4.1.4
django-cors-headers==3.13.0
djangorestframework==3.14.0
gunicorn==20.0.4
psycopg2-binary==2.9.3
python-dotenv==0.21.0
python-environ==0.4.54
pika==0.12.0
requests==2.28.0
whitenoise==6.2.0